backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pydantic==2.7.1
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
supabase==2.4.5
celery==5.4.0
redis==5.0.4
sentry-sdk[fastapi,celery]==2.5.1
pytesseract==0.3.10
Pillow==10.3.0
anthropic==0.28.0
posthog==3.5.0
twilio==9.1.0
python-multipart==0.0.9
stripe==9.9.0
instructor==1.3.4
resend==2.3.0
triggerdotdev==3.0.0
deepgram-sdk==3.4.0
factory-boy==3.3.0
pytest-benchmark==4.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
reportlab==4.2.0
upstash-redis>=1.3.0
